\name{FTICRMS-internal}
\title{Internal FTICRMS objects}
\alias{.benj.hoch}
\alias{.break.clusters}
\alias{.break.dups}
\alias{.centered.coeffs}
\alias{.cluster.matrix}
\alias{.cluster.peaks}
\alias{.loc.maxes}
\alias{.loc.mins}
\alias{.peak.parab}
\alias{.Rsquared}
\description{Internal FTICRMS objects.}
\details{These are not to be called by the user.}
\keyword{internal}