\name{FTICRMS-package}
\alias{FTICRMS-package}
\alias{FTICRMS}
\docType{package}
\title{
Fourier Transform-Ion Cyclotron Resonance Mass Spectrometry (FT-ICR MS) Analysis}
\description{Contains programs for calculating baseline curves, identifying peaks, 
and statistically analyzing FT-ICR MS data.}
\details{
\tabular{ll}{
Package: \tab FTICRMS\cr
Type: \tab Package\cr
Version: \tab 0.6\cr
Date: \tab 2008-09-12\cr
License: \tab GPL-2 \cr
}
This package was developed partially with funding from the NIH Training Program 
in Biomolecular Technology (2-T32-GM08799).
}
\author{
Don Barkauskas

Maintainer: Don Barkauskas (\email{barkda@wald.ucdavis.edu})
}
\keyword{ package }
